+++ .prettierrc
@@ -1,6 +1,7 @@
 {
     "trailingComma": "none",
     "tabWidth": 4,
+    "useTabs": true,
     "semi": true,
     "singleQuote": false,
     "arrowParens": "always",
