+++ src/App.test.ts
@@ -1,0 +1,9 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+import App from "./App";
+
+test("renders learn react link", () => {
+	render(React.createElement(App));
+	// const linkElement = screen.getByText(/learn react/i);
+	// expect(linkElement).toBeInTheDocument();
+});
