--- src/models/UserProfile.js
@@ -1,20 +1,0 @@
-/**
- * User profile factory class.
- */
-export class UserProfile {
-    name = null;
-    mail = null;
-    id = null;
-    isAdmin = false;
-
-    constructor() {
-        this.name = "";
-        this.mail = "";
-        this.id = "";
-        this.isAdmin = false;
-    }
-}
-
-const userProfileInitialState = new UserProfile();
-
-export { userProfileInitialState };
