+++ src/redux/reducers/rootReducer.ts
@@ -1,11 +1,11 @@
 import userProfileReducer from "./userProfileReducer";
+import { UserProfile } from "models";
 
 /**
  * Base structure of our Storage ecosystem.
  */
 const RootReducer = {
-    userProfile: userProfileReducer,
-    someOtherReducer: null
+	userProfile: userProfileReducer
 };
 
 export { RootReducer };
