--- src/redux/selectors/userProfileSelectors.js
@@ -1,1 +1,0 @@
-// TODO: Create data selectors
