+++ tsconfig.json
@@ -3,12 +3,17 @@
         "module": "esnext",
         "target": "esnext",
         "baseUrl": "src",
+        "jsx": "react",
         "paths": {
             "components/*": ["components/*"],
             "services/*": ["services/*"],
             "storage/*": ["storage/*"],
             "router/*": ["components/router/*"]
-        }
+        },
+        "allowJs": true,
+        "allowSyntheticDefaultImports": true,
+        "esModuleInterop": true,
+        "moduleResolution": "node"
     },
     "include": ["src/**/*"]
 }
